+++ Graph-Ting/BinaryTreeCompo.xaml
@@ -1,0 +1,12 @@
+﻿<UserControl x:Class="Graph_Ting.BinaryTreeCompo"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:Graph_Ting"
+             mc:Ignorable="d" 
+             d:DesignHeight="350" d:DesignWidth="500">
+    <Grid>
+        <TreeView BorderThickness="1" x:Name="BinaryTreeView"  />
+    </Grid>
+</UserControl>
