+++ Graph-Ting/MainWindow.xaml
@@ -8,8 +8,10 @@
             <ColumnDefinition Width="1*"/>
             <ColumnDefinition Width="2*"/>
         </Grid.ColumnDefinitions>
-        <ScrollViewer Grid.Column="0" VerticalScrollBarVisibility="Auto" HorizontalScrollBarVisibility="Auto">
-            <TreeView BorderThickness="1" x:Name="BinaryTreeView" FontSize="18" FontFamily="Gadugi"  />
+        <ScrollViewer Grid.Column="0" VerticalScrollBarVisibility="Auto" HorizontalScrollBarVisibility="Auto" 
+                      FontSize="18" FontFamily="Gadugi">
+           
+            <local:BinaryTreeCompo  BorderThickness="1" x:Name="BinaryTreeView"/>
         </ScrollViewer>
         <ScrollViewer Grid.Column="1" x:Name="MainScrollViewer"
                       CanContentScroll="False"
