+++ Graph-Ting/MainWindow.xaml.cs
@@ -10,43 +10,6 @@ namespace Graph_Ting
         public MainWindow()
         {
             InitializeComponent();
-
-            // Sample collection of integers
-            int[] values = { 90,4,56,100,21,455,221,89,6};
-
-            // Create the binary tree and add nodes
-            BinaryTree binaryTree = new BinaryTree();
-            foreach (int value in values)
-            {
-                binaryTree.Insert(value);
-            }
-
-            // Display the binary tree in the TreeView
-            DisplayBinaryTree(binaryTree, BinaryTreeView);
-        }
-
-        private void DisplayBinaryTree(BinaryTree binaryTree, TreeView treeView)
-        {
-            TreeNode root = binaryTree.Root;
-            treeView.Items.Add(CreateTreeViewItem(root));
-        }
-
-        private TreeViewItem CreateTreeViewItem(TreeNode node)
-        {
-            TreeViewItem item = new TreeViewItem();
-            item.Header = node.Value.ToString();
-
-            if (node.Left != null)
-            {
-                item.Items.Add(CreateTreeViewItem(node.Left));
-            }
-
-            if (node.Right != null)
-            {
-                item.Items.Add(CreateTreeViewItem(node.Right));
-            }
-
-            return item;
         }
     }
 }
